+++ Duality/Backend/Audio/INativeAudioSource.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Duality.Backend
+{
+	public interface INativeAudioSource : IDisposable
+	{
+
+	}
+}
