+++ Duality/DualityApp.cs
@@ -11,6 +11,7 @@ using Duality.Backend;
 using Duality.Resources;
 using Duality.Serialization;
 using Duality.Drawing;
+using Duality.Audio;
 using Duality.Cloning;
 using Duality.Input;
