+++ Duality/Resources/Sound.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 using Duality.Editor;
 using Duality.Properties;
+using Duality.Audio;
 
 
 namespace Duality.Resources
@@ -110,7 +111,7 @@ namespace Duality.Resources
 		private	float		pitchFactor		= 1.0f;
 		private	float		fadeOutAt		= 0.0f;
 		private	float		fadeOutTime		= 0.0f;
-		private	SoundType	type			= SoundType.EffectWorld;
+		private	SoundType	type			= SoundType.World;
 		private	List<ContentRef<AudioData>>	audioData	= null;
 
 		/// <summary>
@@ -153,7 +154,7 @@ namespace Duality.Resources
 			set { this.type = value; }
 		}
 		/// <summary>
-		/// [GET / SET] Maximum number of <see cref="Duality.SoundInstance">SoundInstances</see> of this Sound that can
+		/// [GET / SET] Maximum number of <see cref="Duality.Audio.SoundInstance">SoundInstances</see> of this Sound that can
 		/// play simultaneously. If exceeded, any new instances of it are discarded.
 		/// </summary>
 		public int MaxInstances
@@ -178,7 +179,7 @@ namespace Duality.Resources
 			set { this.pitchFactor = value; }
 		}
 		/// <summary>
-		/// [GET / SET] Play time in seconds at which <see cref="Duality.SoundInstance">SoundInstances</see> of this Sound
+		/// [GET / SET] Play time in seconds at which <see cref="Duality.Audio.SoundInstance">SoundInstances</see> of this Sound
 		/// automatically fade out.
 		/// </summary>
 		public float FadeOutAt
